requirements.txt
python-telegram-bot==21.0
python-dotenv>=1.0.1
watchdog==3.0.0
requests==2.31.0
python-dotenv>=1.0.1
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
typing-extensions>=4.5.0
pydantic>=2.0.0
aiohttp>=3.8.0
async-timeout>=4.0.0 
